@@ -7,13 +7,12 @@ class Auth0Auth {
   Auth0Auth._(this.client, this.clientId);
 
   factory Auth0Auth(String clientId, String url, {dynamic bearer}) {
-    assert(clientId != null);
     final _client = new Auth0Client(url, telemetry: telemetry, token: bearer);
     return Auth0Auth._(_client, clientId);
   }
 
-  Future<dynamic> responseHandler(http.Response response) {
-    if (response.statusCode == 200) {
+  Future<dynamic>? responseHandler(http.Response? response) {
+    if (response!.statusCode == 200) {
       return jsonDecode(response.body);
     }
     else if (response.statusCode == 401) {
@@ -102,10 +101,10 @@ class Auth0Auth {
           'client_id': this.clientId,
           'grant_type': 'authorization_code',
         });
-      http.Response res = await this.client.mutate('/oauth/token', payload);
-      return await responseDataHandler(res);
+      http.Response? res = await this.client.mutate('/oauth/token', payload);
+      return await responseDataHandler(res!);
     } catch (e) {
-      throw new Auth0Exeption(description: e);
+      throw new Auth0Exeption(description: e.toString());
     }
   }
 
@@ -130,9 +129,9 @@ class Auth0Auth {
           'client_id': this.clientId,
           'grant_type': 'http://auth0.com/oauth/grant-type/password-realm',
         });
-      http.Response res = await this.client.mutate('/oauth/token', payload);
-      return await responseDataHandler(res);
-    } catch (e) {
+      http.Response? res = await this.client.mutate('/oauth/token', payload);
+      return await responseDataHandler(res!);
+    } on Map catch (e) {
       throw new Auth0Exeption(
           name: e['name'] ?? e['error'],
           description:
@@ -157,10 +156,10 @@ class Auth0Auth {
           'client_id': this.clientId,
           'grant_type': 'refresh_token',
         });
-      http.Response res = await this.client.mutate('/oauth/token', payload);
-      return await responseDataHandler(res);
+      http.Response? res = await this.client.mutate('/oauth/token', payload);
+      return await responseDataHandler(res!);
     } catch (e) {
-      throw new Auth0Exeption(description: e);
+      throw new Auth0Exeption(description: e.toString());
     }
   }
 
@@ -171,10 +170,10 @@ class Auth0Auth {
   //
   Future<dynamic> getUserInfo() async {
     try {
-      http.Response res = await this.client.query('/userinfo');
-      return await responseDataHandler(res);
+      http.Response? res = await this.client.query('/userinfo');
+      return await responseDataHandler(res!);
     } catch (e) {
-      throw new Auth0Exeption(description: e);
+      throw new Auth0Exeption(description: e.toString());
     }
   }
 
@@ -194,8 +193,8 @@ class Auth0Auth {
     return this
         .client
         .mutate('/oauth/revoke', payload)
-        .then((http.Response response) {
-      if (response.statusCode == 200) {
+        .then((http.Response? response) {
+      if (response!.statusCode == 200) {
         return {};
       }
       throw new Auth0Exeption(description: jsonDecode(response.body));
@@ -228,8 +227,8 @@ class Auth0Auth {
     return this
         .client
         .mutate('/dbconnections/change_password', payload)
-        .then((http.Response response) {
-      if (response.statusCode == 200) {
+        .then((http.Response? response) {
+      if (response!.statusCode == 200) {
         return true;
       }
       throw jsonDecode(response.body);
@@ -253,12 +252,12 @@ class Auth0Auth {
       var payload = Map.from(params)..addAll({'client_id': this.clientId});
       if (params['metadata'] != null)
         payload..addAll({'user_metadata': params['metadata']});
-      http.Response res = await this.client.mutate(
+      http.Response? res = await this.client.mutate(
             '/dbconnections/signup',
             payload,
           );
-      return await responseDataHandler(res);
-    } catch (e) {
+      return await responseDataHandler(res!);
+    } on Map<String, dynamic> catch (e) {
       throw new Auth0Exeption(
           name: e['name'], description: e['message'] ?? e['description']);
     }
